server/requirements.txt
yt-dlp
imageio-ffmpeg
Pillow
requests
numpy
torch
transformers
fastapi
uvicorn
pydantic
python-multipart
pip install google-genai 
google-cloud-aiplatform